Update and rename solve.cs to solve-AmirKhaksar.cs
For new contribution policy
